Feat: 그룹 생성 후 화면 이동
UI 및 ViewController 통합
